control through command line
