+++ ses_gen.iml
@@ -13,5 +13,6 @@
     <orderEntry type="library" name="jogl-all-noawt" level="project" />
     <orderEntry type="library" name="com.github.haifengl:smile-core:1.4.0" level="project" />
     <orderEntry type="library" name="math.geom2d:javaGeom:0.11.1" level="project" />
+    <orderEntry type="library" name="com.beust:jcommander:1.72" level="project" />
   </component>
 </module>
